more raycast stuff, organized
